➕[ADD] #196 - 주소 cell과 mapview 같이 넣은 cell 추가
